`furbyte-canvas.js` - constant variables added that have been attacched to some of the accessory pieces for a pet. If the accessory kit is unsucessful then i have a back up...unfortunately the backup is hardcoding images of the pets.
sprites folder has been added to the images folder. it contains `.png`'s of a base model for the cat-pet and a logo designated for the program.

The homepage has changed into a landing page, and the dashboard uses the `furbyte-canvas.js`, still no navigation functions, other than just typing in URL.