`furbyte-canvas.js` has been deleted, learning more about JS, modular sprites and adding JS into the `site,js`
Used JS to draw the first row and first column sprite in the modular sprite sheet. this branch can tecnically be closed and once the default layout is updated to include the dashboard we can continue to work on it. Next the register and login pages need to redirect to dashboard as well.