**CURRENT ERROR- In `RehomeRequests` the database was trying to delete and cascase more than anticipated, i've restricted any deleting of the users.. that's not what I wanted I just want to transfer one pet to another. **(model)
a migration was added and that's when this error was spotted. I will test again but as of right now that is a huge bug that I need to make sure I fixed.

Moving on.. working on the rehome requests. I need a rehome button on the interactions panel maybe??? It can always get moved later.

A couple placeholders are made,.