Seeded Data into the Products Table. The Store modal currently crashes just because there were no pets or products generated, I do need to add-migration and update the database, a few changes have been made thee.
Visually, I needed to have a rehome button on the pet dashboard and get the rehome requests at least ready for me to insert sample data. I wanted to just go through the app and test it that way but I .. need the pets and products.

Loop through the RehomeRequests-- foreach--

there are two templates (P1=>P2) (P1=>OVR), i would like to keep it that way.

UNIT TESTING NEEDS TO BE CONDUCTED CREATE A SEPARATE BRANCH!!!!