Added a funftion LoadStorePartial() so that I can load my modal onclick.. currently I cannot see anything which is causing some stress but.. I'll fix it.
new migration done.

changed the store/index to store index and moved it to the dashboard. I am not sure that was the right move as I now have a bunch of broken components. I think i will need to delete the store controller and readd it if need be.

some od the models were altered but changed right back.