A Store Index was created and the Partial View of _StorePartial was moved to the Store View Folder for the time being.
Altered the IActionResult Methof in the Dashboard Controller so that it can load pet stats instead of the, all of the products and userproducts. I will come back to that and create an inventory for the user??

The Dashboard now retrieves the stats for that pet now I need to develop the logic behind the decay.