Commented out the PetStats DbSet I wasn't sure if it's needed but so far everything seems to be working with it out.
And I removed the PetStats List for now in the StorePageViewModel.. It may return later. But my focus is population.

I have a lot of Indexes or files labeled index and it's confusing so i added a few comments to label them same with the _StorePartialView.

Sincee this is the first time I am doing a modal I reduced it to a smaller chunk until I get it working how I want.

added the storeModal id - to the site.css, so that I can edit that in the future.