Placing a foreach loop within the OverseerDashboard so that it can loop through any rehomerequests that need to be reviewed. this is apart of the CRUD functionality.
A RehomeRequestController has also been added to control th logic of the rehome requests.
The Dashboard button at the very end has also been altered and changed from <button> to <a> so that it will redirect the user to a different form to fill out. That being said, the RehomeRequest Index View was added and BootStrap was used to make form and format this particular index.